+++ pol-core/pol/module/cfgmod.h
@@ -58,7 +58,7 @@ namespace Pol {
 	  EConfigFileRefObjImp( ref_ptr<Core::StoredConfigFile> rcfile );
 	  virtual Bscript::BObjectRef OperSubscript( const Bscript::BObject& obj ) POL_OVERRIDE;
 	  virtual const char* typeOf() const POL_OVERRIDE;
-	  virtual int typeOfInt() const POL_OVERRIDE;
+	  virtual u8 typeOfInt() const POL_OVERRIDE;
 	  virtual Bscript::BObjectImp* copy() const POL_OVERRIDE;
 	};
 
@@ -73,7 +73,7 @@ namespace Pol {
 	  virtual Bscript::BObjectRef get_member( const char* membername ) POL_OVERRIDE;
 	  virtual Bscript::BObjectRef get_member_id( const int id ) POL_OVERRIDE; //id test
 	  virtual const char* typeOf() const POL_OVERRIDE;
-	  virtual int typeOfInt() const POL_OVERRIDE;
+	  virtual u8 typeOfInt() const POL_OVERRIDE;
 	  virtual Bscript::BObjectImp* copy() const POL_OVERRIDE;
 	};
