+++ pol-core/pol/module/partymod.cpp
@@ -55,7 +55,7 @@ namespace Pol {
 	public:
 	  EPartyRefObjImp( Core::PartyRef pref );
 	  virtual const char* typeOf() const POL_OVERRIDE;
-	  virtual int typeOfInt() const POL_OVERRIDE;
+	  virtual u8 typeOfInt() const POL_OVERRIDE;
 	  virtual BObjectImp* copy() const POL_OVERRIDE;
 	  virtual bool isTrue() const POL_OVERRIDE;
 	  virtual bool isEqual( const BObjectImp& objimp ) const POL_OVERRIDE;
@@ -73,7 +73,7 @@ namespace Pol {
 	{
 	  return "PartyRef";
 	}
-	int EPartyRefObjImp::typeOfInt() const
+	u8 EPartyRefObjImp::typeOfInt() const
 	{
 	  return OTPartyRef;
 	}
